Extract test suites into files
