+++ packages/utils/src/__tests__/period/period/testComparePeriods.js
@@ -1,0 +1,145 @@
+import { comparePeriods } from '../../../period/period';
+
+export const testComparePeriods = () => {
+  const testData = [
+    [
+      'year',
+      {
+        '>': [
+          ['2020', '2019'],
+          ['2020', '2020Q3'],
+          ['2020', '202011'],
+          ['2020', '2020W52'],
+          ['2020', '20201230'],
+        ],
+        '=': [
+          ['2020', '2020'],
+          ['2020', '2020Q4'],
+          ['2020', '202012'],
+          ['2020', '20201231'],
+        ],
+        '<': [
+          ['2020', '2021'],
+          ['2020', '2021Q1'],
+          ['2020', '202101'],
+          ['2020', '2021W01'],
+          ['2020', '20210101'],
+        ],
+      },
+    ],
+    [
+      'quarter',
+      {
+        '>': [
+          ['2020Q1', '2019'],
+          ['2020Q1', '2019Q4'],
+          ['2020Q1', '201912'],
+          ['2020Q1', '2019W52'],
+          ['2020Q1', '20191231'],
+        ],
+        '=': [
+          ['2020Q4', '2020'],
+          ['2020Q4', '2020Q4'],
+          ['2020Q4', '202012'],
+          ['2020Q4', '20201231'],
+        ],
+        '<': [
+          ['2020Q4', '2021'],
+          ['2020Q4', '2021Q1'],
+          ['2020Q4', '202101'],
+          ['2020Q4', '2021W01'],
+          ['2020Q4', '20210101'],
+        ],
+      },
+    ],
+    [
+      'month',
+      {
+        '>': [
+          ['202001', '2019'],
+          ['202001', '2019Q4'],
+          ['202001', '201912'],
+          ['202001', '2019W52'],
+          ['202001', '20191231'],
+        ],
+        '=': [
+          ['202012', '2020'],
+          ['202012', '2020Q4'],
+          ['202012', '202012'],
+          ['202012', '20201231'],
+        ],
+        '<': [
+          ['202012', '2021'],
+          ['202012', '2021Q1'],
+          ['202012', '202101'],
+          ['202012', '2021W01'],
+          ['202012', '20210101'],
+        ],
+      },
+    ],
+    [
+      'week',
+      {
+        '>': [
+          ['2020W01', '2019'],
+          ['2020W01', '2019Q4'],
+          ['2020W01', '201912'],
+          ['2020W01', '2019W52'],
+          ['2020W01', '20191231'],
+        ],
+        '=': [
+          ['2020W53', '2020W53'],
+          ['2020W52', '20201227'],
+          ['2020W53', '20210103'],
+        ],
+        '<': [
+          ['2020W52', '2021'],
+          ['2020W52', '2021Q1'],
+          ['2020W52', '202101'],
+          ['2020W52', '2021W01'],
+          ['2020W52', '20201228'],
+          ['2020W53', '20210301'],
+        ],
+      },
+    ],
+
+    [
+      'day',
+      {
+        '>': [
+          ['20200101', '2019'],
+          ['20200101', '2019Q4'],
+          ['20200101', '201912'],
+          ['20191230', '2019W52'],
+          ['20200101', '20191231'],
+        ],
+        '=': [
+          ['20201231', '2020'],
+          ['20201231', '202012'],
+          ['20201231', '2020Q4'],
+          ['20201227', '2020W52'],
+          ['20210103', '2020W53'],
+        ],
+        '<': [
+          ['20201231', '2021'],
+          ['20201231', '2021Q1'],
+          ['20201231', '202101'],
+          ['20201231', '2021W01'],
+          ['20201231', '20210101'],
+        ],
+      },
+    ],
+  ];
+
+  it.each(testData)('%s', (_, testCaseData) => {
+    testCaseData['>'].forEach(([periodA, periodB]) => {
+      expect(comparePeriods(periodA, periodB)).toBeGreaterThan(0);
+    });
+    testCaseData['='].forEach(([periodA, periodB]) => {
+      expect(comparePeriods(periodA, periodB)).toBe(0);
+    });
+    testCaseData['<'].forEach(([periodA, periodB]) => {
+      expect(comparePeriods(periodA, periodB)).toBeLessThan(0);
+    });
+  });
+};
